First stage support for 'halftone'. Wrapper code for pgimag, pggray done.
This is going to be a messy operation. Will require multiple commits for
working the correct header transformations (wip has some good code for this
that will eventually need to be ported to python). For temporary status,
we can hard code in a simple transformation matrix to get a 1-1 correspondance
between pixel position and grid position. PGIMAG and PGGRAY is not a simple
wrapper since they have complex structure (but a few lines can be eliminated
since they have almost identical function definitions. NOTE: The wrapper code
specifically checks to make sure the transformation matrix has 6 elements
since it appears that PGPLOT just assumes such a matrix.